Home Credit Default Risk/credit_default_app/requirements.txt
streamlit
pandas
numpy
matplotlib
mysql-connector-python
pandas-profiling
streamlit-pandas-profiling
pydantic
pydantic-settings
python-dotenv